+++ src/test/converter2/issues/gh2552.js
@@ -6,12 +6,12 @@
  */
 
 // TS 5.5 @import comments
-/** @import * as ts from "typescript" */
+/** @import ts from "typescript" */
 
 /**
  * This is an awesome module.
  * @module good-module
  */
 
-/** @import * as ts2 from "typescript" */
+/** @import ts2 from "typescript" */
 export const something = 1;